requirements/base.txt
django==1.9.8
djangorestframework==3.4.0
django-rest-swagger==2.0.2
psycopg2==2.6.2
PyYaml==3.11
